Rename study/id to study/studyid
하위 엔드포인트의 commentid 와의 구별을 위해 study/id 를 study/studyid 로 수정
