+++ src/routes/study/studyid/index.ts
